Rename "models" packadge to "entities"
