+++ src/main/java/org/example/animalchipization/dto/account/AccountDtoOut.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
-import org.example.animalchipization.models.Account;
+import org.example.animalchipization.entities.Account;
 
 /**
  * Represents account data transfer object
