+++ src/main/java/org/example/animalchipization/dto/animal/AnimalDtoOut.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 import org.example.animalchipization.enums.AnimalGender;
 import org.example.animalchipization.enums.AnimalLifeStatus;
 
-import org.example.animalchipization.models.Animal;
+import org.example.animalchipization.entities.Animal;
 
 
 import java.time.Instant;
