+++ src/main/java/org/example/animalchipization/dto/location/LocationDtoIn.java
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
-import org.example.animalchipization.models.Location;
+import org.example.animalchipization.entities.Location;
 
 /**
  * Represents location data transfer object
