+++ src/main/java/org/example/animalchipization/entities/AnimalType.java
@@ -1,4 +1,4 @@
-package org.example.animalchipization.models;
+package org.example.animalchipization.entities;
 
 import jakarta.persistence.*;
 import lombok.*;
