+++ src/main/java/org/example/animalchipization/enums/AnimalGender.java
@@ -1,6 +1,6 @@
 package org.example.animalchipization.enums;
 
-import org.example.animalchipization.models.Animal;
+import org.example.animalchipization.entities.Animal;
 
 /**
  * Represents possible gender values for animal.
